backend/requirements.txt
# Django Core
Django==5.1.5
djangorestframework==3.15.2

# Database
psycopg2-binary==2.9.10

# Environment variables
python-decouple==3.8

# CORS headers for frontend integration
django-cors-headers==4.6.0

# Image handling
Pillow==11.0.0

# API Documentation
drf-spectacular==0.28.0

# Async tasks (Redis)
redis==5.2.1
celery==5.4.0

# Timezone support
pytz==2024.2

# Development tools
django-extensions==3.2.3
